refactor: modal
